--- src/mocks/handlers.ts
@@ -1,40 +1,0 @@
-import { http, HttpResponse } from "msw";
-import products from "./data/products.json";
-import cart from "./data/cart.json";
-import { CART_URL, PRODUCT_URL } from "../constants/endpoint";
-import filterProductList from "../../test/utils/filterProductList";
-import sortProductList from "../../test/utils/sortProductList";
-import { FilterType, SortType } from "../types";
-
-export const handlers = [
-	http.get(PRODUCT_URL, ({ request }) => {
-		const url = new URL(request.url);
-		const category = url.searchParams.get("category") as FilterType;
-		const sort = url.searchParams.get("sort");
-
-		let resultProducts = products;
-		if (category) resultProducts = filterProductList(resultProducts, category);
-		if (sort) resultProducts = sortProductList(resultProducts, sort.split(",")[1] as SortType);
-
-		return HttpResponse.json({ content: resultProducts });
-	}),
-
-	http.get(CART_URL, () => {
-		return HttpResponse.json({ content: cart });
-	}),
-
-	http.post(CART_URL, async ({ request }) => {
-		const body = await request.json();
-		const { productId, quantity } = body as { productId: number; quantity: number };
-
-		if (!productId || quantity < 1) {
-			return HttpResponse.error();
-		}
-
-		return HttpResponse.json({ ok: true, status: 201 });
-	}),
-
-	http.delete(`${CART_URL}/:id`, async () => {
-		return HttpResponse.json({ ok: true, status: 201 });
-	}),
-];
