+++ src/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+import { cartHandler } from "./cartHandler";
+import { productHandler } from "./productHandler";
+
+export const handlers = [...cartHandler, ...productHandler];
